pyproject.toml
[tool.poetry]
name = "notes-django-ninja"
version = "0.1.0"
description = "Проект notes django-ninja + Axios + Alpine.js"
authors = ["MrLerich <93153582+MrLerich@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "notes_django_ninja"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
django-ninja = "^0.20.0"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.5"
psycopg2 = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
